+++ torchvision/models/detection/__init__.py
@@ -1,0 +1,1 @@
+from .mask_rcnn import maskrcnn_resnet50_fpn, MaskRCNN
